backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
alembic
aiogram
python-dotenv
python-jose[cryptography]
passlib[bcrypt]
pgvector
httpx
pydantic-settings
slowapi
celery[redis]
redis
sentence-transformers
exa_py
gunicorn
openai
kafka-python==2.0.2
